refactor: 에러 메시지 변경 및 클래스 분리, 코드 가독성 높이기
