+++ core/src/main/kotlin/com/wespot/auth/dto/apple/ApplePublicKey.kt
@@ -1,0 +1,10 @@
+package com.wespot.auth.dto.apple
+
+data class ApplePublicKey(
+    val kty: String,
+    val kid: String,
+    val use: String,
+    val alg: String,
+    val n: String,
+    val e: String
+)
