+++ core/src/main/kotlin/com/wespot/auth/service/SocialAuthServiceFactory.kt
@@ -2,13 +2,14 @@ package com.wespot.auth.service
 
 import com.wespot.user.SocialType
 import org.springframework.stereotype.Component
+import org.springframework.stereotype.Service
 
-@Component
+@Service
 class SocialAuthServiceFactory(
     private val services: List<SocialAuthService>
 ) {
     fun getService(socialType: SocialType): SocialAuthService {
         return services.find { it.isSupport(socialType) }
-            ?: throw IllegalArgumentException("Unsupported social type: $socialType")
+            ?: throw IllegalArgumentException("해당 소셜로그인을 지원하지 않습니다. social type: $socialType")
     }
 }
